♻️ refactor: 기본페이지 뒤로가기 미설정으로 인한 폴더구조 재조정
