+++ src/app/feed/(list)/page.tsx
@@ -1,9 +1,5 @@
 import FeedContent from "@/components/feed/feedContent";
 
 export default function FeedPage() {
-  return (
-    <>
-      <FeedContent />
-    </>
-  );
+  return <FeedContent />;
 }
