+++ src/app/mypage/(edit)/edit/page.tsx
@@ -1,9 +1,5 @@
 import ProfileItemEdit from "@/components/form/mypage/profileItemEdit";
 
 export default function MypageEdit() {
-  return (
-    <>
-      <ProfileItemEdit />
-    </>
-  );
+  return <ProfileItemEdit />;
 }
