+++ src/app/mypage/(edit)/layout.tsx
@@ -1,5 +1,5 @@
 import Navbar from "@/components/Navbar";
-import "../../styles/globals.css";
+import "../../../styles/globals.css";
 import { ChevronLeft } from "lucide-react";
 
 export default function NavbarLayout({
