+++ src/app/mypage/(view)/layout.tsx
@@ -1,6 +1,5 @@
 import Navbar from "@/components/Navbar";
 import "../../../styles/globals.css";
-import { ChevronLeft } from "lucide-react";
 
 export default function NavbarLayout({
   children,
@@ -10,10 +9,10 @@ export default function NavbarLayout({
   return (
     <div>
       <div className="relative w-full px-4 py-5">
-        <button className="absolute -translate-y-1/2 cursor-pointer left-4 top-1/2">
+        {/* <button className="absolute -translate-y-1/2 cursor-pointer left-4 top-1/2">
           <ChevronLeft />
-        </button>
-        <p className="text-center">살펴보기</p>
+        </button> */}
+        <p className="text-center">마이페이지</p>
       </div>
       <div className="relative w-full px-4 pb-25">{children}</div>
       <Navbar />
