+++ src/app/mypage/(view)/page.tsx
